+++ swe/src/main/java/com/kwu/swe/domain/lecture_location/repository/LectureLocationRepository.java
@@ -1,6 +1,6 @@
-package com.kwu.swe.domain.lecture.repository;
+package com.kwu.swe.domain.lecture_location.repository;
 
-import com.kwu.swe.domain.lecture.entity.LectureLocation;
+import com.kwu.swe.domain.lecture_location.entity.LectureLocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LectureLocationRepository extends JpaRepository<LectureLocation, Long> {
