+++ swe/src/main/java/com/kwu/swe/domain/lecture_schedule/repository/LectureScheduleRepository.java
@@ -1,6 +1,6 @@
-package com.kwu.swe.domain.lecture.repository;
+package com.kwu.swe.domain.lecture_schedule.repository;
 
-import com.kwu.swe.domain.lecture.entity.LectureSchedule;
+import com.kwu.swe.domain.lecture_schedule.entity.LectureSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LectureScheduleRepository extends JpaRepository<LectureSchedule, Long> {
